- Dissolve now works on very long samples.

+++ src/filters/edit.h
@@ -159,7 +159,7 @@ private:
   PClip child2;
   const int overlap;
   int video_fade_start, video_fade_end;
-  int audio_fade_start, audio_fade_end;
+  __int64 audio_fade_start, audio_fade_end;
   BYTE* audbuffer;
   int audbufsize;
   void EnsureBuffer(int minsize);
